fix some warning and re-format

+++ src/PufferfishIndex.cpp
@@ -2,15 +2,13 @@
 #include <iostream>
 
 #include "CLI/Timer.hpp"
+#include "CanonicalKmerIterator.hpp"
 #include "PufferFS.hpp"
 #include "PufferfishIndex.hpp"
-#include "CanonicalKmerIterator.hpp"
 #include "cereal/archives/binary.hpp"
 #include "cereal/archives/json.hpp"
 
 #include "jellyfish/mer_dna.hpp"
-
-
 
 PufferfishIndex::PufferfishIndex() {}
 
@@ -129,14 +127,13 @@ uint32_t PufferfishIndex::contigID(CanonicalKmer& mer) {
   return std::numeric_limits<uint32_t>::max();
 }
 
-
-//this function would store the part of the
-//contig that is required
-//if hitFW is true then we will clip
-//the part towards the end of the contig
+// this function would store the part of the
+// contig that is required
+// if hitFW is true then we will clip
+// the part towards the end of the contig
 /*
-void PufferfishIndex::getContigSeq(bool hitFW, uint64_t globalPos, int clipLen ,std::string& contigStr){
-  auto queryPos = kit->second ;
+void PufferfishIndex::getContigSeq(bool hitFW, uint64_t globalPos, int clipLen
+,std::string& contigStr){ auto queryPos = kit->second ;
 
   uint64_t globalPos = phits.globalPos_ ;
   uint64_t twoGlobalPos = globalPos << 1 ;
@@ -147,31 +144,34 @@ void PufferfishIndex::getContigSeq(bool hitFW, uint64_t globalPos, int clipLen ,
 /*
 //TODO does the orientation of the contig matter while chopping sequence ?
 //not needed
-void PufferfishIndex::getRawSeq(util::ProjectedHits& phits, CanonicalKmerIterator& kit,  std::string& contigStr, int readLen){
+void PufferfishIndex::getRawSeq(util::ProjectedHits& phits,
+CanonicalKmerIterator& kit,  std::string& contigStr, int readLen){
   //auto queryPos = kit->second ;
 
   //auto remainingContigLen = phits.contigLen_ - (phits.contigPos_ + k_) ;
   //auto remainingReadLen = readLen - (kit->second + k_ ) ;
-    
+
     uint64_t globalPos = phits.globalPos_ ;
     //uint64_t twoGlobalPos = globalPos << 1 ;
 
     //index of the contig
     auto rank = contigRank_(globalPos) ;
     //start position of this contig
-    uint64_t sp = (rank == 0) ? 0 : static_cast<uint64_t>(contigSelect_(rank)) + 1;
+    uint64_t sp = (rank == 0) ? 0 : static_cast<uint64_t>(contigSelect_(rank)) +
+1;
 
     //clip from both side for now
     //go to right by read length
     //if contigLen < 2*readLen then
-    //get back the entire contig 
+    //get back the entire contig
     uint64_t clipLen = 2*readLen ;
     uint64_t clipStart = sp ;
     if(phits.contigLen_ < clipLen){
       clipLen = phits.contigLen_ ;
     }else{
       clipStart = std::max(sp, globalPos - readLen) ;
-      clipLen = std::min(readLen,static_cast<int>(globalPos - clipStart)) + std::min(phits.contigPos_ + 101, phits.contigLen_ - phits.contigPos_) ;
+      clipLen = std::min(readLen,static_cast<int>(globalPos - clipStart)) +
+std::min(phits.contigPos_ + 101, phits.contigLen_ - phits.contigPos_) ;
     }
     uint64_t twoSp = clipStart << 1 ;
     auto numOfKmers = clipLen / k_ ;
@@ -184,7 +184,7 @@ void PufferfishIndex::getRawSeq(util::ProjectedHits& phits, CanonicalKmerIterato
       contigStr += mer.to_str() ;
       i++ ;
     }
-      
+
     //TODO check boundary
     //residual kmer we might clip the wrong end
     if(resLen > 0){
@@ -201,7 +201,8 @@ void PufferfishIndex::getRawSeq(util::ProjectedHits& phits, CanonicalKmerIterato
 }
 */
 
-auto PufferfishIndex::getRefPos(CanonicalKmer& mer, util::QueryCache& qc) -> util::ProjectedHits {
+auto PufferfishIndex::getRefPos(CanonicalKmer& mer, util::QueryCache& qc)
+    -> util::ProjectedHits {
   using IterT = std::vector<util::Position>::iterator;
   auto km = mer.getCanonicalWord();
   size_t res = hash_raw_->lookup(km);
@@ -220,14 +221,14 @@ auto PufferfishIndex::getRefPos(CanonicalKmer& mer, util::QueryCache& qc) -> uti
       // start position of this contig
       uint64_t sp = 0;
       uint64_t contigEnd = 0;
-      if ( rank == qc.prevRank ) {
+      if (rank == qc.prevRank) {
         sp = qc.contigStart;
         contigEnd = qc.contigEnd;
       } else {
         sp = (rank == 0) ? 0 : static_cast<uint64_t>(contigSelect_(rank)) + 1;
         contigEnd = contigSelect_(rank + 1);
         qc.prevRank = rank;
-        qc.contigStart= sp;
+        qc.contigStart = sp;
         qc.contigEnd = contigEnd;
       }
 
@@ -243,15 +244,30 @@ auto PufferfishIndex::getRefPos(CanonicalKmer& mer, util::QueryCache& qc) -> uti
       // how the k-mer hits the contig (true if k-mer in fwd orientation, false
       // otherwise)
       bool hitFW = (keq == KmerMatchType::IDENTITY_MATCH);
-      return {rank, pos, relPos, hitFW, static_cast<uint32_t>(clen), k_,
+      return {static_cast<uint32_t>(rank),
+              pos,
+              relPos,
+              hitFW,
+              static_cast<uint32_t>(clen),
+              k_,
               core::range<IterT>{pvec.begin(), pvec.end()}};
     } else {
-      return {std::numeric_limits<uint32_t>::max(), std::numeric_limits<uint64_t>::max(), std::numeric_limits<uint32_t>::max(), true, 0, k_,
+      return {std::numeric_limits<uint32_t>::max(),
+              std::numeric_limits<uint64_t>::max(),
+              std::numeric_limits<uint32_t>::max(),
+              true,
+              0,
+              k_,
               core::range<IterT>{}};
     }
   }
 
-  return {std::numeric_limits<uint32_t>::max(), std::numeric_limits<uint64_t>::max(), std::numeric_limits<uint32_t>::max(), true, 0, k_,
+  return {std::numeric_limits<uint32_t>::max(),
+          std::numeric_limits<uint64_t>::max(),
+          std::numeric_limits<uint32_t>::max(),
+          true,
+          0,
+          k_,
           core::range<IterT>{}};
 }
 
@@ -272,7 +288,8 @@ auto PufferfishIndex::getRefPos(CanonicalKmer& mer) -> util::ProjectedHits {
       // the reference information in the contig table
       auto& pvec = contigTable_[rank];
       // start position of this contig
-      uint64_t sp = (rank == 0) ? 0 : static_cast<uint64_t>(contigSelect_(rank)) + 1;
+      uint64_t sp =
+          (rank == 0) ? 0 : static_cast<uint64_t>(contigSelect_(rank)) + 1;
       uint64_t contigEnd = contigSelect_(rank + 1);
 
       // relative offset of this k-mer in the contig
@@ -287,15 +304,30 @@ auto PufferfishIndex::getRefPos(CanonicalKmer& mer) -> util::ProjectedHits {
       // how the k-mer hits the contig (true if k-mer in fwd orientation, false
       // otherwise)
       bool hitFW = (keq == KmerMatchType::IDENTITY_MATCH);
-      return {rank,  pos, relPos, hitFW, static_cast<uint32_t>(clen), k_,
+      return {static_cast<uint32_t>(rank),
+              pos,
+              relPos,
+              hitFW,
+              static_cast<uint32_t>(clen),
+              k_,
               core::range<IterT>{pvec.begin(), pvec.end()}};
     } else {
-      return {std::numeric_limits<uint32_t>::max(),  std::numeric_limits<uint64_t>::max(), std::numeric_limits<uint32_t>::max(), true, 0, k_,
+      return {std::numeric_limits<uint32_t>::max(),
+              std::numeric_limits<uint64_t>::max(),
+              std::numeric_limits<uint32_t>::max(),
+              true,
+              0,
+              k_,
               core::range<IterT>{}};
     }
   }
 
-  return {std::numeric_limits<uint32_t>::max(),  std::numeric_limits<uint64_t>::max(), std::numeric_limits<uint32_t>::max(), true, 0, k_,
+  return {std::numeric_limits<uint32_t>::max(),
+          std::numeric_limits<uint64_t>::max(),
+          std::numeric_limits<uint32_t>::max(),
+          true,
+          0,
+          k_,
           core::range<IterT>{}};
 }
 
@@ -313,6 +345,6 @@ const std::string& PufferfishIndex::refName(uint64_t refRank) {
   return refNames_[refRank];
 }
 
-const std::vector<std::string>& PufferfishIndex::getRefNames(){
-    return refNames_ ;
+const std::vector<std::string>& PufferfishIndex::getRefNames() {
+  return refNames_;
 }
